+++ src/darker/import_sorting.py
@@ -1,4 +1,15 @@
 import logging
+import sys
+from pathlib import Path
+from typing import Optional
+
+from black import find_project_root
+
+if sys.version_info >= (3, 8):
+    from typing import TypedDict
+else:
+    from typing_extensions import TypedDict
+
 
 try:
     import isort
@@ -8,19 +19,30 @@ except ImportError:
 logger = logging.getLogger(__name__)
 
 
-def apply_isort(content: str) -> str:
-    isort_config_kwargs = dict(
-        multi_line_output=3,
-        include_trailing_comma=True,
-        force_grid_wrap=0,
-        use_parentheses=True,
-        line_length=88,
-        quiet=True,
-    )
+class IsortArgs(TypedDict, total=False):
+    line_length: int
+    settings_file: str
+    settings_path: str
+
+
+def apply_isort(
+    content: str,
+    src: Optional[Path] = None,
+    config: Optional[str] = None,
+    line_length: Optional[int] = None,
+) -> str:
+    isort_args = IsortArgs()
+    if src and not config:
+        isort_args["settings_path"] = str(find_project_root((str(src),)))
+    if config:
+        isort_args["settings_file"] = config
+    if line_length:
+        isort_args["line_length"] = line_length
+
     logger.debug(
         "isort.code(code=..., {})".format(
-            ", ".join(f"{k}={v!r}" for k, v in isort_config_kwargs.items())
+            ", ".join(f"{k}={v!r}" for k, v in isort_args.items())
         )
     )
-    result: str = isort.code(code=content, **isort_config_kwargs)
+    result: str = isort.code(code=content, **isort_args)
     return result