+++ CHANGES.rst
@@ -5,6 +5,7 @@
 - Feature: Add support for ``-l``/``--line-length`` and ``-S``/``--skip-string-normalization``
 - Feature: ``--diff`` outputs a diff for each file on standard output
 - Feature: Require ``isort`` >= 5.0.1 and be compatible with it.
+- Feature: Allow to configure ``isort`` through pyproject.toml
 
 
 0.2.0 / 2020-03-11